back/requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
